Issue #15: Fixed problem with wrist markers 5-8 always off.
The bit mask used to cycle through the markers was a short, i.e.
limited to 16 bits (16 markers). I changed it to a long to accommodate
all 20 GRIP markers.

+++ GripMMI/GripMMIData.cpp
@@ -63,7 +63,7 @@ int GripMMIDesktop::GetGripRT( void ) {
 	int bytes_read;
 	int return_code;
 	static unsigned short previousTMCounter = 0;
-	unsigned short bit = 0;
+	unsigned long bit = 0;
 	int retry_count;
 
 	EPMTelemetryPacket		packet;
@@ -347,7 +347,7 @@ int GripMMIDesktop::GetLatestGripHK( GripHealthAndStatusInfo *hk ) {
 	int bytes_read;
 	int return_code;
 	static unsigned short previousTMCounter = 0;
-	unsigned short bit = 0;
+	unsigned long bit = 0;
 	int retry_count;
 
 	EPMTelemetryPacket packet;
